pyproject.toml
[project]
name = "vllm-grpc"
version = "0.0.0"
requires-python = ">=3.12,<3.13"
description = "Protobuf/gRPC frontend for vLLM"
readme = "README.md"
dependencies = []

[tool.uv.workspace]
members = ["packages/gen", "packages/proxy", "packages/frontend", "packages/client", "tools/benchmark"]

[dependency-groups]
# vllm-metal 0.2.0 is installed here; vllm is added at runtime via --with vllm
# because vllm 0.20.0 has CUDA-only deps with no macOS wheels (nvidia-cudnn-frontend).
# The install.sh builds vllm from source to work around this — we use --with vllm
# to match the existing project pattern, letting uv find the macOS-compatible version.
investigation = [
    "vllm-metal @ https://github.com/vllm-project/vllm-metal/releases/download/v0.2.0-20260430-132616/vllm_metal-0.2.0-cp312-cp312-macosx_11_0_arm64.whl",
]
dev = [
    "grpcio-tools>=1.65",
    "modal>=0.73",
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "grpc-stubs>=1.53",
    "openai>=1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--import-mode=importlib"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"

[[tool.mypy.overrides]]
module = [
    "vllm_grpc.*",
    "grpc_tools.*",
    "vllm_grpc.v1.chat_pb2",
    "vllm_grpc.v1.chat_pb2_grpc",
]
ignore_errors = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["vllm", "vllm.*", "transformers", "transformers.*", "torch", "torch.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["huggingface_hub", "huggingface_hub.*"]
ignore_missing_imports = true
